+++ CHANGELOG.md
@@ -22,6 +22,7 @@
     - Provides more ways to register models. In particular, there is now the option to register a model's metadata via a "model info file".
     - !19 improves the implementation of `utopya models copy`, becoming more language-agnostic and extendable.
 - !4 and !6 add a validated project registry, which keeps track of utopya projects (which in turn contain models) and frameworks: two abstraction levels that play a role in determining the configuration hierarchy.
+    - !24 improves the registry framework and the CLI to better handle corrupt project (or model) registry files.
 - !5 adds changes that allow a proper [outsourcing of utopya from Utopia](https://gitlab.com/utopia-project/utopia/-/merge_requests/277):
     - Extends the project registration CLI with `--require-matching-names` option.
     - Allows to specify custom plot configuration pools via the meta-configuration.